reformating project code

+++ C++/Problem001.cpp
@@ -1,7 +1,3 @@
-#include "pch.h"
-#include "Problems.h"
-
-
 /*
 =================
 =   Problem 1   =	Multiples of 3 and 5
@@ -13,10 +9,15 @@ we get 3, 5, 6 and 9. The sum of these multiples is 23.
 Find the sum of all the multiples of 3 or 5 below 1000.
 */
 
+#include "pch.h"
+
+#include <iostream>
+
+
 // Check if number is divisible by 3 or 5 and add to the sum
-unsigned int problem1(unsigned int lim) {
+int problem1(int lim) {
 	unsigned int ans = 0;
-	for (size_t i = 0; i < lim; ++i) {
+	for (int i = 0; i < lim; ++i) {
 		if (i % 3 == 0 || i % 5 == 0) {
 			ans += i;
 		}
